+++ packages/core/transpiler/view-generator/src/NodeGenerators/ForGenerator.ts
@@ -36,8 +36,8 @@ export default class ForGenerator extends BaseGenerator {
   /**
    * @View
    * ${dlNodeName} = new ForNode(${array}, ${depNum}, ${array}.map(${item} => ${key}))
-   * ${dlNodeName}.addNodeFunc((${item}, $updateMap, $key) => {
-   *   $updateMap.set($key, (changed, $item) => {
+   * ${dlNodeName}.addNodeFunc((${item}, $updateArr, $idx) => {
+   *   $updateArr[$idx] = (changed, $item) => {
    *      ${item} = $item
    *      {$updateStatements}
    *   })
@@ -61,27 +61,26 @@ export default class ForGenerator extends BaseGenerator {
     childStatements.unshift(
       ...this.declareNodes(nodeIdx),
       this.t.expressionStatement(
-        this.t.callExpression(
+        this.t.assignmentExpression(
+          "=",
           this.t.memberExpression(
-            this.t.identifier("$updateMap"),
-            this.t.identifier("set")
+            this.t.identifier("$updateArr"),
+            this.t.identifier("$idx"),
+            true
           ),
-          [
-            this.t.identifier("$key"),
-            this.t.arrowFunctionExpression(
-              [...this.updateParams, this.t.identifier("$item")],
-              this.t.blockStatement([
-                this.t.expressionStatement(
-                  this.t.assignmentExpression(
-                    "=",
-                    item,
-                    this.t.identifier("$item")
-                  )
-                ),
-                ...this.geneUpdateBody(updateStatements).body,
-              ])
-            ),
-          ]
+          this.t.arrowFunctionExpression(
+            [...this.updateParams, this.t.identifier("$item")],
+            this.t.blockStatement([
+              this.t.expressionStatement(
+                this.t.assignmentExpression(
+                  "=",
+                  item,
+                  this.t.identifier("$item")
+                )
+              ),
+              ...this.geneUpdateBody(updateStatements).body,
+            ])
+          )
         )
       )
     )
@@ -111,8 +110,8 @@ export default class ForGenerator extends BaseGenerator {
             this.t.arrowFunctionExpression(
               [
                 item as any,
-                this.t.identifier("$updateMap"),
-                this.t.identifier("$key"),
+                this.t.identifier("$updateArr"),
+                this.t.identifier("$idx"),
               ],
               this.t.blockStatement(childStatements)
             ),